warning resolution

+++ serverless_functions/aws/java/cpu_test/src/main/java/cpu_test/Handler.java
@@ -35,7 +35,7 @@ public class Handler implements RequestStreamHandler {
 		// extract number to factorize if in request
 		long n;
 		if (event.containsKey("n")) {
-			n = (long)Math.round((double)event.get("n"));
+			n = Math.round((double)event.get("n"));
 		} else {
 			n = 71950288374236L;
 		}
